requirements.txt
openai-whisper
pydub
h5py
tgt
torchtext
h5py
validators
tqdm
numpy
argparse
requests
colorama
tabulate
transformers
scikit-learn
Flask
Flask-Reuploaded
flask-executor
webvtt-py